feat: delete room implement layer
