--- src/main/kotlin/smalltalk/backend/application/implement/room/Implement.kt
@@ -1,9 +1,0 @@
-package smalltalk.backend.application.implement.room
-
-import org.springframework.stereotype.Component
-
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-@MustBeDocumented
-@Component
-annotation class Implement(val value: String = "")
